img_vector_skill/requirements.txt
fastapi==0.104.1
pydantic==2.5.2
pydantic_core==2.14.5
httpx==0.25.2
python-dotenv==1.0.0
uvicorn==0.24.0.post1
azure-storage-blob==12.19.0
requests==2.31.0
openai==1.3.5